requirements.txt
pandas
selenium
beautifulsoup4
PyYAML
loguru
numpy
matplotlib
requests
psycopg2-binary
SQLAlchemy
dotenv